python_consumer/requirements.txt
boto3
numpy
pandas
faker
python-dotenv